revise place order

+++ src/main/java/com/webdesign/bestsell/controller/OrderController.java
@@ -10,6 +10,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
@@ -39,53 +40,44 @@ public class OrderController {
     /**
      * place order
      * localhost:8080/pri/order/place_order
-     * format:
-     * {
-     *     "userId":2,
-     *     "productId":2
-     * }
-     * @param order
      * @return
      */
     @PostMapping("place_order")
     @Transactional
-    public JsonData placeOrder(@RequestBody Order order) {
+    public JsonData placeOrder() {
 
-        int userId = order.getUserId();
-
-        order.setCreateTime(new Date());
-        int productId = order.getProductId();
-
-        // if there is items left
-        Product product = productService.getProductById(productId);
-        if (product.getStock() <= 0) {
-            return JsonData.buildError("no item left!");
-        }
-
-        // if item is from current user
-        List<Product> products = productService.getProductByUserId(userId);
-        for (Product p: products) {
-            if (p.getId() == productId) {
-                return JsonData.buildError("It's your item!");
-            }
-        }
-
-        // remove items from user cart
+        List<Order> orderList = new ArrayList<>();
+        Order order = null;
+        int userId = 1;
         List<Cart> cartList = userService.getCartByUserId(userId);
         for (Cart cart: cartList) {
+            int productId = cart.getProductId();
+            order = new Order();
+            order.setUserId(userId);
+            order.setProductId(productId);
+            order.setCreateTime(new Date());
+
+            // if there is items left
+            Product product = productService.getProductById(productId);
+            if (product.getStock() <= 0) {
+                return JsonData.buildError("no item left!");
+            }
+
+            // update items stock
+            product.setStock(product.getStock() - 1);
+            int row = productService.updateStock(product);
+            if (row <= 0) {
+                return JsonData.buildError("error when attempt to get item");
+            }
+
+            int result = userService.placeOrder(order);
+            if (result <= 0) { return JsonData.buildError("error when place order");}
+
             userService.deleteItemFromCart(cart.getId());
+            orderList.add(order);
         }
 
-        // update items stock
-        product.setStock(product.getStock() - 1);
-        System.out.println(product.getStock());
-        int row = productService.updateStock(product);
-        if (row <= 0) {
-            return JsonData.buildError("error when attempt to get item");
-        }
-        row = userService.placeOrder(order);
-
-        return row > 0 ? JsonData.buildSuccess(order): JsonData.buildError("success place");
+        return JsonData.buildSuccess(orderList);
     }
 
     /**
